Removed the dependency on the source MAC when adding a tunnel
The ARP request which is issued by the gateway before sending the
received inner packet to the client can not be responded to if
the client and the gateway are on different subnets.

This was previously solved by having the MAC address of the client
known when adding the tunnel, and then adding an OpenFlow rule which
would respond to the ARP request when issued.

In order to remove the dependency on knowing the client's L2 MAC
address, the ARP response rule was replaced with a rule that changes
the source IP in the ARP request, to be the tunnel's destination IP.
Since the two need sides of the tunnel to be able to communicate,
this ensures that the ARP requests will be correctly responded to by
the client when issued by the gateway.

Signed-off-by: Gavi Teitz <gavi@mellanox.com>